feat(condo): DOMA-8507 add JEST_MOCKS_ENABLED env variable to make sure that we use mocked clients when start condo within separated process
