+++ .github/workflows/nodejs.condo.cypress.yml
@@ -74,7 +74,7 @@ jobs:
           
           yarn workspace @app/condo node ./bin/generate-initial-data.js
 
-          echo 'ADDRESS_SERVICE_CLIENT_MODE=fake' >> .env
+          echo 'JEST_MOCKS_ENABLED=true' >> .env
           echo 'DISABLE_CAPTCHA=true' >> .env
   
           yarn workspace @app/condo start &
