+++ apps/condo/bin/prepare.js
@@ -4,7 +4,7 @@ async function updateAppEnvAddressSuggestionConfig (serviceName) {
     const addressServiceUrl = await getAppServerUrl('address-service')
     await updateAppEnvFile(serviceName, 'ADDRESS_SERVICE_URL', addressServiceUrl)
     // NOTE: address-service must be in fake mode by default to pass tests. Later we should prepare address-service instead to work locally!
-    await updateAppEnvFile(serviceName, 'ADDRESS_SERVICE_CLIENT_MODE', 'fake')
+    await updateAppEnvFile(serviceName, 'JEST_MOCKS_ENABLED', 'true')
 }
 
 async function main () {
